// src/interface/canvas-setting-info.interface.ts
/**
 * canvas 畫布設定
 *
 * @export
 * @interface CanvasSettingInfo
 */
export interface CanvasSettingInfo {

  /**
   * 寬度
   *
   * @type {number}
   * @memberof CanvasSettingInfo
   */
  width: number;

  /**
   * 高度
   *
   * @type {number}
   * @memberof CanvasSettingInfo
   */
  height: number;

}